pyproject.toml
[project]
name = "playwright-demo"
version = "0.1.0"
description = "BDD test automation with Playwright and pytest-bdd"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "allure-pytest-bdd==2.16.0",
    "playwright==1.59.0",
    "pytest==9.0.3",
    "pytest-bdd==8.1.0",
    "pytest-html==4.2.0",
    "pytest-playwright==0.7.2",
    "pytest-rerunfailures==16.1",
    "pytest-xdist==3.8.0",
]

[tool.pytest.ini_options]
# Step files can use scenarios("file.feature") without path prefixes.
bdd_features_base_dir = "tests/features"
testpaths = ["tests/steps"]
addopts = [
    "-v",
    "--tb=short",
    "--html=reports/html/report.html",
    "--self-contained-html",
    "--alluredir=reports/allure-results",
    "--clean-alluredir",
    "--output=reports/playwright-artifacts",
    "--tracing=on",
    "--video=on",
    "--screenshot=only-on-failure",
]
markers = [
    "smoke: critical path checks to run before every deployment",
    "regression: full regression suite",
    "orangehrm: OrangeHRM demo application tests",
    "sauce: SauceDemo application tests",
    "auth: authentication tests",
    "pim: OrangeHRM PIM module tests",
    "leave: OrangeHRM Leave module tests",
    "inventory: inventory page tests",
    "cart: cart page tests",
    "checkout: checkout flow tests",
    "e2e: end-to-end tests",
    "pta: PracticeTestAutomation site tests",
    "table: PracticeTestAutomation table tests",
    "learning: examples that teach pytest-bdd syntax without using a browser",
    "parsers: parser examples for string, parse, regex, and cfparse steps",
    "datatable: data table examples for pytest-bdd",
    "login: authentication tests",
    "negative: negative-path scenarios",
    "smoke_pta: Smoke testing for PTA",
    "smoke_sd: Smoke testing for SauceDemo",
]
